requirements.txt
psycopg2-binary==2.9.9
pyTelegramBotAPI==4.23.1
openai==1.12.0
python-dotenv==1.0.0

Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
django-filter==23.5
djangorestframework-simplejwt==5.3.1
Pillow==10.1.0
python-dotenv==1.0.0